@BrunoViniciusB Componentes adicionados: CardCuidador, ContactCuidador, DateTimeTextButton e ImageCuidador. Screens atualizadas: ChatListScreen, SearchScreen, SearchResultScreen
