+++ app/src/main/java/elder/ly/mobile/MainActivity.kt
@@ -11,8 +11,10 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import elder.ly.mobile.ui.screens.chatlist.ChatListScreen
 import elder.ly.mobile.ui.screens.search.SearchScreen
 import elder.ly.mobile.ui.screens.signupstep1.SignUpStep1Screen
+import elder.ly.mobile.ui.screens.signupstep2.SignUpStep2Screen
 import elder.ly.mobile.ui.theme.MobileTheme
 
 class MainActivity : ComponentActivity() {
@@ -31,7 +33,7 @@ class MainActivity : ComponentActivity() {
 
 @Composable
 fun App(modifier: Modifier = Modifier) {
-    SearchScreen()
+    ChatListScreen()
 }
 
 @Preview(showBackground = true)
