+++ app/src/main/java/elder/ly/mobile/ui/screens/chatlist/ChatListScreen.kt
@@ -3,54 +3,60 @@ package elder.ly.mobile.ui.screens.chatlist
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
-import androidx.compose.material3.HorizontalDivider
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Icon
 import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import elder.ly.mobile.R
 import elder.ly.mobile.R.drawable.ic_pesquisar
-import elder.ly.mobile.ui.components.Feature
+import elder.ly.mobile.ui.components.Contacts
+import elder.ly.mobile.ui.components.DefaultTextInput
 import elder.ly.mobile.ui.components.NavBar
 import elder.ly.mobile.ui.theme.MobileTheme
-import elder.ly.mobile.ui.theme.secondaryContainerLight
 import elder.ly.mobile.ui.theme.tertiaryContainerLight
+import kotlin.math.max
 
 @Composable
 fun ChatListScreen() {
+    var searchText by remember { mutableStateOf("") }
+
     Column(
         modifier = Modifier
-            .background(Color.White)
             .fillMaxSize()
     ) {
         Column(
             modifier = Modifier
-                .background(Color.White)
                 .padding(top = 16.dp, bottom = 0.dp, start = 16.dp, end = 16.dp)
         ) {
-            SearchChat()
+            SearchChat(
+                searchText = searchText,
+                onSearchTextChange = { newText ->
+                    searchText = newText
+                }
+            )
         }
 
         LazyColumn(
@@ -67,93 +73,43 @@ fun ChatListScreen() {
 }
 
 @Composable
-fun SearchChat() {
-    Box(
+fun SearchChat(searchText: String, onSearchTextChange: (String) -> Unit) {
+    Row(
         modifier = Modifier
             .fillMaxWidth()
-            .shadow(
-                elevation = 4.dp,
-                shape = RoundedCornerShape(16.dp),
-                clip = false
-            )
-            .background(
-                color = Color.White,
-                shape = RoundedCornerShape(8.dp)
-            )
             .border(
                 width = 1.dp,
                 color = tertiaryContainerLight,
                 shape = RoundedCornerShape(8.dp)
-            )
-            .padding(8.dp)
+            ),
+        horizontalArrangement = Arrangement.Center,
+        verticalAlignment = Alignment.CenterVertically
     ) {
+
         Image(
             painter = painterResource(id = ic_pesquisar),
             contentDescription = "Ícone de pesquisa",
             modifier = Modifier
                 .size(48.dp)
-                .align(alignment = Alignment.CenterStart)
-                .padding(start = 8.dp)
+                .padding(start = 12.dp),
         )
-    }
-}
 
-@Composable
-fun Contacts() {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-    ) {
-        Row(
+        OutlinedTextField(
+            value = searchText,
+            onValueChange = { onSearchTextChange(it) },
+            placeholder = { Text(text = "Pesquisar", color = Color.Gray) },
             modifier = Modifier
-                .padding(start = 16.dp, end = 16.dp, bottom = 8.dp, top = 8.dp)
-                .height(88.dp)
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.ic_launcher_background),
-                contentDescription = "Foto de perfil do cuidador",
-                modifier = Modifier
-                    .size(48.dp)
-                    .align(Alignment.CenterVertically),
-                Alignment.CenterStart
-            )
-
-            Column(
-                modifier = Modifier
-                    .padding(start = 8.dp)
-                    .align(Alignment.CenterVertically)
-
-            ) {
-                Text(text = "Vila Matilde", color = secondaryContainerLight)
-                Text(text = "Maria Antonieta", fontSize = 18.sp)
-                LazyRow {
-                    items(listOf("Fraldas", "Bingo", "Medicação")) { featureText ->
-                        Feature(text = featureText)
-                    }
-                }
-            }
-            Column (
-                modifier = Modifier
-                    .weight(1f)
-                    .align(Alignment.CenterVertically),
-                horizontalAlignment = Alignment.End
-            ){
-                Arrow()
-            }
-        }
-        HorizontalDivider(modifier = Modifier.fillMaxWidth(), thickness = 1.dp, color = Color.Gray)
-    }
-}
-
-@Composable
-fun Arrow() {
-    Column{
-        Icon(
-            imageVector = Icons.AutoMirrored.Filled.KeyboardArrowRight,
-            contentDescription = "Arrow Forward",
-            tint = Color(0xFF343643),
-            modifier = Modifier
-                .size(48.dp)
+                .fillMaxWidth(),
+            shape = RoundedCornerShape(8.dp),
+            colors = OutlinedTextFieldDefaults.colors(
+                unfocusedBorderColor = Color.Transparent,
+                focusedBorderColor = Color.Transparent,
+                cursorColor = Color.Gray
+            ),
+            keyboardOptions = KeyboardOptions.Default.copy(
+                imeAction = ImeAction.Search
+            ),
+            maxLines = 1
         )
     }
 }
